Prevent automake from failing when we have no ChangeLog.
By default, automake tries to insure that a project adheres to
the GNU Coding Standards.  It will fail if there is no ChangeLog.
The --foreign keyword tells automake that a project is not trying to
follow the GNU Coding Standards.  Missing ChangeLogs and the like
are no longer fatal errors.
This patch adds --foreign to both automake invocations.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>