started cleaning up configure.ac
This update to our build system is based on results I received when
running autoscan.  Also, I found that we were using obsolete macros and
some obsolete forms of calling macros, so these have been updated.

The build.sh script is no longer needed due to being able to run
autoreconf, and the version.sh file is no longer needed since the
version number is set in the call to AC_INIT.

Signed-off-by: William Hubbs <w.d.hubbs@gmail.com>