Use glib consistently in command handlers.
The command handlers in module_utils.c intermixed malloc and g_malloc.
In some cases, we couldn't even predict which allocator was used
for a function's return value.
After this patch, all of the handlers return g_malloc'ed data.
In module_main.c, we free it with g_free, rather than xfree.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>