Stop using xfree in the audio subsystem.
This patch gets rid of the use of xfree in the audio subsystem.

Along the way, I found a bug in the alsa code.  A pointer was not being
set to NULL after memory was freed,
thus leaving a dangling reference.
Also, alsa_play could leak memory under an error condition.