Use glib allocation in the server.
Now, glib is used for all memory allocation in the server, with a handful
of exceptions.  A few variables are modified by external functions
(like getline), and they need to be managed by malloc and free.
The functions from src/common/ still use strdup, so memory returned by
those also needs to be freed with libc's free.
Finally, includepath is eventually freed by dotconf, so it needs to be
allocated using libc's strdup.