Use glib allocation in the modules.
All of the allocation in the various output modules is now done
exclusively by glib, with a handful of exceptions.
The getline function requires that we use malloc and free to manage
its second parameter (the line buffer).  So we still use libc's malloc
and free to manage those.  There aren't many of them.