Properly indent header files.
The following construct is used in public header files.  By default,
it causes Lindent to add an unwanted level of indentation to the file.

extern "C" {
...
}

This construct needs to be surrounded with control comments, so that
Lindent will ignore it:
/* *INDENT-OFF* */
/* *INDENT-ON* */

I added the control comments, and re-indented the public headers.