Fix segfault when compiled against libao 1.0.0.
Version 1.0.0 of libao adds a new member to the ao_sample_format
structure.  It's called "matrix".  This member was never initialized,
because our ao_sample_format structure is allocated on the stack.
The matrix member should be NULL if we don't care about the channel
matrix.  We don't care about it, so this patch sets the contents of
the ao_sample_format structure to zero.