Use otts_getline instead of getline.
Several copies of the getline function were scattered throughout the
code. I got rid of the duplicates, and I moved an implementation into
the convenience library.

Also, I rewrote getline slightly, so that it complies with glibc's
implementation.
Notably, glibc's implementation will write the line to a buffer that
was pre-allocated by the user, resizing the buffer with realloc when
necessary.

Finally, I changed the name of the function to otts_getline, and I
changed all call sites so that they call it, rather than getline.