Autospawn mechanism fine-tuning
The autospawn script has been eliminated.  Now, the server program
takes a --spawn option, which should be used when autospawn is requested.
When --spawn is given, openttsd will look into its configuration,
and if it cannot find a ''DisableAutoSpawn'' parameter,
then it will start, otherwise it will exit with an error value of 1.

The ~/.speech-dispatcher/ directory and configuration files are no longer
required to exist before openttsd is started. If they don't exist,
openttsd will use the system configuration from /etc/opentts/
and will create this directory including the necessary log and pid subdirs.

Running openttsd from /etc/init.d/ is no longer recommended, and there
is no direct support for it.
The current arrangement is however general enough, so that you can
setup an /etc/init.d/openttsd script if you wish, by specifying the pid,
log and conf paths on the command line on startup.

There is no way from the client to determine whether the user wants to
start a user session or the system openttsd.
So far, this has been determined by uid==0, but
this is not a good method. Openttsd should generally never run as root.
Now, autospawn always starts openttsd as the current user.

Another reason for the elimination of the special autospawn shell script is that
it needed to perform a duplicate determination of the correct
configuration file to check for DisableAutoSpawn.

This change was imported from the Speech Dispatcher project. See
commit ab0cd4ec432e77f952703b8e76134fadaf9e04e1 from that project.
The original author was Hynek Hanke <hanke@brailcom.org>.