Do not include the common library in libopentts.
This was a very bad mistake on my part, because libopentts was exporting
internal functions from the common library.

Instead, we should just have two copies of the otts_getline function:
one for libopentts, and a second for everything else.
The copy in libopentts is static, so it is not exported.