Move spd_audio.c and spd_audio.h to the modules subdirectory.
These are only used by the modules, and there is no need to put them in
a public shared library.
They have a low memory footprint, because all of the real work is done
in the plugins.

Also, rename them to audio.c and audio.h.

This patch also fixes a stray comment in src/audio/alsa.c, so that it
refers to the correct file.