Add constants for reporting errors.
The functions str2voice, str2priority, str2recogn, str2punct, and
str2spell now return special "error" constants when the string
argument is unrecognized.
We use these in dotconf callbacks and the function that parses
module settings.

The dotconf callbacks for DefaultCapLetRecognition,
DefaultPunctuationMode, DefaultVoice, and DefaultPriority
were also rewritten to produce an error message which is more descriptive
than "Invalid parameter in configuration file."
For example, we will now see
"Invalid value foo for the DefaultCapLetRecognition parameter in the
configuration file", when the user supplies an incorrect value such
as foo.