fix includes of libopentts.h and opentts_types.h in the say client and api
The C api source was including both opentts_types.h and libopentts.h
which is not necessary because libopentts.h includes opentts_types.h.
In the say client, options.h was including opentts_types.h which was
causing it to be included twice in say.c for the same reason.