remove duplicate include directive in openttsd.conf
There were originally two include directives in this file, one with an
absolute path and the other with a relative path.  They were both
including the same files since dotconf supports relative paths.
Removing the include directive with the absolute path makes it possible
to include this file as is without auto generating it.