Replace many references to Speech Dispatcher.
This commit is based on a patch that I received from Andrei Kholodnyi.
That patch was edited significantly.

References to Speech Dispatcher remain un-replaced in the Czech
documentation.  Andrei's patch did replace those, but I think we
should leave the Czech documentation alone until we have a translator.

modules/festival_client.c contains references to speech-dispatcher, but
those must stay, because the festival client relies on
festival-freebsoft-utils, which provides the speech-dispatcher module.