Fix two memory leaks in the parse_set function.
The parse_set function uses a macro called TEST_CMD to test all the
subcommands of the set command.  It has a side-effect: it frees
the subcommand if it is the one that we were looking for.
We no longer need it.
The old code used !strcmp when testing for the rate and pause_context
subcommands, rather than TEST_CMD.  So every time we received
set self rate x
or
set self pause_context x
we leaked the memory containing the string "rate".

(Thanks to Andrei Kholodnyi for pointing out pause_context).