Fix a nasty off-by-one bug in the Festival module.
A for loop in festival_client.c is supposed to iterate over a string,
copying everything that is not a parenthesis.
The comparison in the loop's test should be <, not <=.
Also, the statement that will append a NUL byte to the end of the copy
was wrong.  It was appending the number 0, rather than the character \0.