Fix an error-handling clause in alsa_play.
My patch to fix the suspend bug was flawed.
It called the ERROR_EXIT macro.  That macro has side-effects; it closes
the alsa device and frees a pointer named track_volume.samples.
However, track_volume.samples is not initialized at this point in the
function.  It should not be freed.
I changed the ERROR_EXIT call to "return -1;".  That is what it should
have been in the first place.

Also, MSG(1, "...") should have been ERR("..."), since the ERR
macro is used for reporting errors.