The function flite_set_voice should be a no-op.
We only support one voice, and that voice is established in
module_init.  The voice data is managed by the flite library.
We should not try to free it.
The previous version of flite_set_module simply freed the voice
data and re-registered the voice that was set in module_init.