build fixes for the audio subsystem
This patch fixes the following issues:
If there are no audio subsystems enabled, the preloaded symbols list was
not defined.  This has been fixed by adding the "-dlopen force" option
to the linker command line.
Also, the configure script now generates a warning if there are no audio
subsystems detected or enabled.