Properly handle autospawning in the Python library.
If autospawning is requested, then we first try to connect to the
server without spawning.  If that fails, we try the connection again.
The server will be spawned on the second try.

Also, the default for the autospawn argument of the SSIPClient
constructor is now True, rather than None.

I removed useless checks for the existence of the spawn module, since
it is always available.