possible memory leaks
The functions set_numeric_parameter, do_audio and do_loglevel have a
situation in which the memory allocated by the function otts_getline is not
released. Within a loop they test if some variables is equal to NULL and if
the condition is met they return to the beginning of the loop, without
release the memory allocated to the variable line. The first thing in the
loop is to set the variable line to NULL.

Signed-off-by: Luke Yelavich <themuso@opentts.org>