Wait for the auto-spawned server process to terminate.
When the Python library auto-spawns a server process, it should wait
for that process to terminate.  This should happen almost immediately,
because the auto-spawned process calls the daemon() function to
start the server in the background.

This change was imported from the Speech Dispatcher project.  See
commit 0f3d8e4981df40aaa0a973e5ce46214729883540 from that repository.
The author of that change was Hynek Hanke <hanke@brailcom.org>.