Separate signal handling into its own thread.
Two signals change the state of the running server.
SIGHUP causes it to reread its configuration file, and SIGUSR1 causes
it to reload dead output modules.
The problem with these two actions is that they were not thread-safe.
The "speaking" thread could be left with a dangling pointer after
one of them completed.

We now put signal handling in its own thread.  When one of these
state-changing signals is caught, the speaking thread is stopped and
restarted.

Thanks to Andrei Kholodnyi for valuable feedback regarding this patch.