check for a module INIT response just right after it was sent
When the server sent the INIT command to an output module, it was not
checking the status code returned by the module.
As of now, the server detects a failed initialization as soon as it
can, by examining the status code of the reply.

Edited by Chris, as follows.  Re-indented the load_output_module
function, and added extra descriptive text to the commit message.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>