Refactor the logging code.
Both the server and the modules now use the same function for logging.
It is named log_msg.
The server's custom logging facility uses a function named log_msg2.
All of these logging functions are defined together in one file.

Note that the audio plugins are a special case.  They cannot directly
use log_msg, because they are dynamically loaded.  In order for them
to use the same log levels, log files, and mutex as the module in
which they run, we need to pass in the module's logging function by
pointer.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>