Rename the libao module to ao.
After a recent patch, the name of the libao module started causing
confusion for the linker.  For some odd reason, it wasn't linking the
plugin libao.so against the system library libao.so.
So I changed the name of the plugin from libao to ao.
Either the new name (ao) and the old name (libao) can be used in the
AudioOutputMethod setting of openttsd.conf, so users need not update
their configuration files.