Fix a buffer overflow in src/server/openttsd.c.
When the file descriptor of a new client socket is greater than or
equal to status.num_fds - 1, we reallocate the array of client sockets.
The size used in the g_realloc call is status.num_fds * 2, and this is
wrong, because the file descriptor of the client socket can be greater
than or equal to status.num_fds * 2.