Check for loaded modules and default module in configure().
We need to check for loaded modules in the configure() function,
rather than the init() function, because the set of available modules
can change every time we reload the configuration file.

Also, if the default output module could not be found, we need to warn
the user and set the default module to the first one that is available.
This latter check came from a patch originally submitted by
Andrei Kholodnyi.