Open an initial connection to PulseAudio in pulse_open.
The previous implementation of pulse_open just allocated data.  It did
not actually try to connect to the PulseAudio server.
We should make a preliminary connection, so that if it fails, we
can possibly try to use another audio backend.

Since we may need to reconnect in pulse_play, the code that makes a
connection was moved into a helper function, pulse_open_helper.

This change is based on commit 25facedf4dd13e41fb3eabe72f58bff0cb0eb86f
from the Speech Dispatcher repository.  The author of that commit was
Hynek Hanke <hanke@brailcom.org>.