make building several synthesizer modules optional
The dummy module will always be built since there needs to be a fallback
module in case nothing else works.

The cicero and festival modules will not be built by default, but can be
enabled using the --with-cicero and --with-festival command line switches
to configure.

The generic module will be built by default, but can be disabled with
the --without-generic command line switch to configure.

These patches and this commit message have been modified by William
Hubbs.

Signed-off-by: William Hubbs <w.d.hubbs@gmail.com>

added conditional modules support to the modules configs

Signed-off-by: William Hubbs <w.d.hubbs@gmail.com>