cleanup execs in src/server/module.c
We don't need to check if exec failed in module.c if it returns an
error has taken place, so we can just log the error and exit.

Also, get rid of the cfg and arg1 variables.  They are only used in
the execlp calls, and they duplicate existing data.

(Based on work from Trevor Saunders.  Commit message edited by Chris.)

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>