close logfiles on exec
this issue has been around for a while, the close on exec bit was never
set on the log files, which meant that modules would have fd that
pointed to the server's log file etc.  We fix this by using fcntl to set
the close on exec bit after opening the file.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>