add system service mode
System service mode is activated with -y or --system-service on the
openttsd command line.

System mode must be run as root, and it will change to another user
and group, which are configurable when running ./configure.  Also, it is
possible to configure where the log files, pid file and unix socket are
placed.

The following options have been added to the configure script:

--with-opentts-user (user for the system service, the default is opentts)

--with-opentts-group (group for the system service, the default is opentts)

--with-system-log-path (path for log file for system service, the
default is /var/log/opentts)

--with-system-pid-path (path to pid file for system service, the default
is /var/run)

--with-system-socket-path (path to the socket for the system service,
the default is /var/run/openttsd)

I would like to thank Chris Brannon for his assistance with some of the
details in this feature.