Document the system-service feature.
We need to document it in two places.  First, packagers need to know
how to create the infrastructure for system-service mode.  That is
described in README.packagers.
Second, users need a description of the steps necessary to start the
daemon as a system service.  The process is described in the OpenTTS
manual.