Fix config/desktop/Makefile.am, for make distcheck.
We don't distribute openttsd.desktop, so it needs to be listed in one
of the nodist_ variables.
Next, the name of the EXTRA_DIST variable needs to be completely
uppercase.  It was mis-spelled as extra_DIST.
Finally, since openttsd.desktop is built by make, it needs to be listed
in CLEANFILES, rather than DISTCLEANFILES.