Move .desktop files to data/desktop.
Openttsd.desktop and other .desktop files are data not configuration.
The rational for this is that these files are used be other applications
to get information about a program, to autostart it
or put it in the main menu, for instance.
So have nothing to do with opentts internal configuration.