Python: refactor SSIPClient to eliminate circular reference.
This patch was submited to speech-dispatcher by
Christopher Brannon <cmbrannon79@gmail.com>.
Merging was done by hand.

The callback-handling portion of SSIPClient is now factored into a
class named _CallbackHandler.  _CallbackHandler implements the __call__
method, so it is callable.  An SSIPClient shares its _CallbackHandler
object with its _SSIPConnection object, so _SSIPConnection no longer
contains a reference to a bound method of SSIPClient.
The circular reference is gone.

Signed-off-by: Andrei Kholodnyi <Andrei.Kholodnyi@gmail.com>